Fix restart container

+++ container_start.sh
@@ -1,15 +1,13 @@
 #!/bin/bash
-cleanup() {
-    echo "Removing X server lock files."
-    killall Xvfb
-    killall x11vnc
-    rm -f /tmp/.X1-lock
-    rm -f /tmp/.X11-unix/X1
-}
-trap cleanup EXIT
+if [ -f /tmp/.X1-lock ] || [ -f /tmp/.X11-unix/X1 ]; then
+  echo "Removing X server lock files."
+  killall Xvfb
+  killall x11vnc
+  rm -f /tmp/.X1-lock
+  rm -f /tmp/.X11-unix/X1
+fi
 
 echo "starting X server and VNC display"
-
 touch ~/.Xauthority
 Xvfb :1 -screen 0 1920x1080x24 -noreset & 
 sleep 5
